Fix ambiguous sin/abs for Xcode 26 TestFlight archive
Color(white:) wants Double while filing positions are CGFloat, so
abs(sin(f.p.y * 0.05)) is ambiguous between CoreGraphics and
Foundation. Pin IronFilings to Foundation.sin(Double). Same mixed
CGFloat/Double sites in Chladni, balloon sway, and face specular
get explicit types so the next archive does not trip the same way.

+++ Interactions/Experiments/FaceSpecular/FaceSpecularView.swift
@@ -98,7 +98,7 @@ private enum FaceSpecularRenderer {
                 let y = center.y + i
                 streak.move(to: CGPoint(x: center.x - radius * 0.94, y: y))
                 streak.addLine(to: CGPoint(x: center.x + radius * 0.94, y: y + i * 0.016))
-                inner.stroke(streak, with: .color(Color.white.opacity(0.038 + abs(i / radius) * 0.012)), lineWidth: 0.9)
+                inner.stroke(streak, with: .color(Color.white.opacity(0.038 + abs(Double(i / radius)) * 0.012)), lineWidth: 0.9)
             }
             let hx = center.x + offset.width * radius * 0.55
             let hy = center.y + offset.height * radius * 0.55
